Introduce internal RFC3986 helper class that gathers the URIDecoder static methods as well as some of HttpUtils URI related methods.
