+++ vertx-core/src/main/java/io/vertx/core/impl/verticle/PackageHelper.java
@@ -22,7 +22,7 @@ import java.util.Enumeration;
 import java.util.List;
 import java.util.jar.JarEntry;
 
-import static io.vertx.core.internal.net.URIDecoder.decodeURIComponent;
+import static io.vertx.core.internal.net.RFC3986.decodeURIComponent;
 
 /**
  * @author Janne Hietam&auml;ki
